paralegal_compiler
Module common

Modules

ast

templates

verify_scope

vis

This visitor is incomplete. For now we only use it to edit AST nodes and the rule numbers and to count variables, so that’s what the traversals reach. If you need more functionality you have to add it.

Structs

Policy

Enums

PolicyScope

Functions

count_references_to_variable