paralegal_compiler::common
Module vis

This visitor is incomplete. For now we only use it to edit AST nodes and the rule numbers and to count variables, so that’s what the traversals reach. If you need more functionality you have to add it.

Traits

Visit

VisitMut

Functions

super_visit_ast_node

super_visit_ast_node_mut

super_visit_clause

super_visit_clause_intro

super_visit_clause_intro_mut

super_visit_clause_mut

super_visit_definition

super_visit_definition_mut

super_visit_joined_nodes

super_visit_joined_nodes_mut

super_visit_only_via

super_visit_only_via_mut

super_visit_policy

super_visit_policy_mut

super_visit_relation

super_visit_relation_mut

super_visit_variable_intro

super_visit_variable_intro_mut